code cleanup, small bug fixes
