+++ accounts/signals.py
@@ -5,7 +5,6 @@ from django.contrib.auth.management import create_permissions
 
 from project import settings
 
-
 # def post_migrate_create_organization(sender, **kwargs):
 #    for app_config in apps.get_app_configs():
 #        create_permissions(app_config, apps=apps, verbosity=0)
