+++ exp/views/analytics.py
@@ -1,10 +1,12 @@
 import datetime
 import json
 from collections import Counter, defaultdict
+
 from django.core.paginator import Paginator
 from django.core.serializers.json import DjangoJSONEncoder
 from django.views import generic
 from guardian.mixins import PermissionRequiredMixin as ObjectPermissionRequiredMixin
+
 from accounts.models import Child, User
 from exp.utils import RESPONSE_PAGE_SIZE
 from exp.views.mixins import ExperimenterLoginRequiredMixin
