+++ studies/permissions.py
@@ -69,103 +69,17 @@ road" in terms of properly segmenting permissions, or represent things
 that simply aren't modeled in the database.
 """
 
+from collections import namedtuple
 from enum import Enum
 
 # Upgrade to python 3.8 for cached_property
 # from functools import cached_property
 from typing import NamedTuple, Tuple
 
-# from django.contrib.auth.models import Group
-# from guardian.shortcuts import assign_perm
+from django.contrib.auth.models import Group, Permission
+from django.contrib.contenttypes.models import ContentType
 
-
-class _PermissionName(Enum):
-    # Model-level
-    ADD = "add"
-    CHANGE = "change"
-    DELETE = "delete"
-    VIEW = "view"
-    # Field-level
-    EDIT = "edit"
-    READ = "read"
-    ASSOCIATE = "associate"
-    DISSOCIATE = "dissociate"
-
-
-class _PermissionParts(NamedTuple):
-    permission: str
-    model_name: str
-    relationship_path: Tuple[str, ...]
-    target_fields: Tuple[str, ...]
-
-
-class _PermissionMeta(NamedTuple):
-    """Immutable, object-oriented wrapper for Study and Lab permissions.
-
-    We exploit the fact that private fields are not returned when iterating
-    over tuples.
-
-    Properties:
-        codename: a string codename. See module docstring for conventions.
-        description: a string description.
-    """
-
-    codename: str
-    description: str
-
-    @property
-    def parts(self) -> _PermissionParts:
-        perm_and_model, *field_vector = self.codename.split("__")
-        permission, model_name = perm_and_model.split("_")
-
-        if field_vector:
-            final_field_spec = field_vector.pop()
-            target_fields = tuple(final_field_spec.strip("[]").split("|"))
-        else:
-            target_fields = ()
-
-        return _PermissionParts(
-            permission=permission,
-            model_name=model_name,
-            relationship_path=tuple(field_vector),
-            target_fields=target_fields,
-        )
-
-    @property
-    def permission(self) -> str:
-        return self.parts.permission
-
-    @property
-    def model_name(self) -> str:
-        return self.parts.model_name
-
-    @property
-    def relationship_path(self) -> Tuple[str, ...]:
-        return self.parts.relationship_path
-
-    @property
-    def target_fields(self) -> Tuple[str, ...]:
-        return self.parts.target_fields
-
-    @classmethod
-    def from_parts(
-        cls,
-        permission: str,
-        model_name: str,
-        relationship_path: Tuple[str, ...],
-        target_fields: Tuple[str, ...],
-        description: str,
-    ):
-        codename = f"{permission}_{model_name}"
-        if relationship_path:
-            codename += f"__{'__'.join(relationship_path)}"
-        if target_fields:
-            if len(target_fields) > 1:
-                codename += f"__[{'|'.join(target_fields)}]"
-            else:
-                codename += f"__{target_fields[0]}"
-
-        return cls(codename, description)
+from project.permissions import _PermissionMeta, _PermissionName, _PermissionParts
 
 
 def create_lab_version(study_permission: _PermissionMeta):
@@ -282,7 +196,6 @@ class LabPermission(_PermissionMeta, Enum):
     ]
 
 
-# Groups (mainly for use in migrations)
 class LabGroup(set, Enum):
     READ = {LabPermission.READ_STUDY_DETAILS, LabPermission.READ_STUDY_PREVIEW_DATA}
     ADMIN = {
@@ -356,28 +269,11 @@ class StudyGroup(set, Enum):
     }
 
 
-def create_groups_for_instance(
-    model_instance, group_enum, group_class, perm_class, group_object_permission_model
-):
-    uuid_segment = str(model_instance.uuid)[:7]
-    object_name = model_instance._meta.object_name
-    unique_group_tag = (
-        f"{object_name} :: {model_instance.name[:7]}... ({uuid_segment}...)"
-    )
-
-    for group_spec in group_enum:
-        # Group name is going to be something like "READ :: Lab :: MIT (0235dfa...)
-        group_name = f"{group_spec.name} :: {unique_group_tag}"
-        group = group_class.objects.create(name=group_name)
-
-        for permission_meta in group_spec.value:
-            permission = perm_class.objects.get(codename=permission_meta.codename)
-
-            group_object_permission_model.objects.create(
-                content_object=model_instance, permission=permission, group=group
-            )
-            group.save()
-
-        setattr(model_instance, f"{group_spec.name.lower()}_group", group)
-
-    model_instance.save()
+# Permissions for Lookit-wide groups affecting all studies/labs.
+# Minimal now but a placeholder for where
+# we can flesh out permissions for e.g. RAs working on tech support vs. recruitment.
+# In studies app because perms do apply to Study, Lab models specifically; Lookit-wide
+# groups are defined under accounts and perms relating to all users may be there also in the
+# future.
+# Codenames used here are NOT directly in line with model/study level perms used elsewhere
+# for ease of avoiding conflicts and removing these permissions later.
